Move binder psi elements to binder package
