+++ src/main/java/org/purescript/psi/binder/PSNullBinderImpl.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.binder
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSNullBinderImpl(node: ASTNode) : PSPsiElement(node)
