+++ src/main/java/org/purescript/psi/declaration/PSValueDeclaration.kt
@@ -4,10 +4,10 @@ import com.intellij.lang.ASTNode
 import com.intellij.navigation.ItemPresentation
 import com.intellij.psi.*
 import org.purescript.features.DocCommentOwner
-import org.purescript.psi.PSBinderImpl
+import org.purescript.psi.binder.PSBinderImpl
 import org.purescript.psi.PSPsiElement
 import org.purescript.psi.PSPsiFactory
-import org.purescript.psi.PSVarBinderImpl
+import org.purescript.psi.binder.PSVarBinderImpl
 import org.purescript.psi.expression.PSExpressionWhere
 import org.purescript.psi.name.PSIdentifier
 import javax.swing.Icon
