+++ src/test/java/org/purescript/accessors.kt
@@ -8,7 +8,7 @@ import org.purescript.psi.PSForeignDataDeclaration
 import org.purescript.psi.PSForeignValueDeclaration
 import org.purescript.psi.PSModule
 import org.purescript.psi.declaration.PSValueDeclaration
-import org.purescript.psi.PSVarBinderImpl
+import org.purescript.psi.binder.PSVarBinderImpl
 import org.purescript.psi.classes.PSClassConstraint
 import org.purescript.psi.classes.PSClassDeclaration
 import org.purescript.psi.classes.PSClassMember
